funkcni mining module pro spectatora ktery pridava goldy na zaklade poctu aktualne tezenych elementu

+++ Application/Core/Players/Player.cs
@@ -185,7 +185,7 @@ namespace Orbit.Core.Players
                 if (lastScoreValue != Data.Score || lastGoldValue != Data.Gold)
                 {
                     lastScoreValue = Data.Score;
-                    lastGoldValue = Data.Gold;
+                    lastGoldValue = (int) Data.Gold;
                     SendScoreAndGoldUpdate();
                 }
                 scoreUpdateTimer = 0;
